+++ despesas-backend-api-net-core.XUnit/Business/Implementations/UsuarioBusinessImplTest.cs
@@ -1,5 +1,4 @@
 ﻿using despesas_backend_api_net_core.Business.Implementations;
-using despesas_backend_api_net_core.Domain.Entities;
 using despesas_backend_api_net_core.Infrastructure.Data.EntityConfig;
 using despesas_backend_api_net_core.Infrastructure.Data.Repositories.Generic;
 
@@ -10,18 +9,16 @@ namespace Test.XUnit.Business.Implementations
         private readonly Mock<IRepositorio<Usuario>> _repositorioMock;
         private readonly UsuarioBusinessImpl _usuarioBusiness;
         private List<Usuario> _usuarios;
-        private List<UsuarioVM> _usuarioVMs;
 
         public UsuarioBusinessImplTest()
         {
             _repositorioMock = new Mock<IRepositorio<Usuario>>();
             _usuarioBusiness = new UsuarioBusinessImpl(_repositorioMock.Object);
             _usuarios = UsuarioFaker.Usuarios();
-            _usuarioVMs = UsuarioFaker.UsuariosVMs();
         }
 
         [Fact]
-        public void Create_ReturnsParsedUsuarioVM()
+        public void Create_Should_Returns_Parsed_UsuarioVM()
         {
             // Arrange
             var usuario = _usuarios.First();
@@ -32,18 +29,21 @@ namespace Test.XUnit.Business.Implementations
             var result = _usuarioBusiness.Create(new UsuarioMap().Parse(usuario));
 
             // Assert
+            Assert.NotNull(result);
+            Assert.IsType<UsuarioVM>(result);
             Assert.Equal(usuario.Id, result.Id);
+            _repositorioMock.Verify(repo => repo.Insert(It.IsAny<Usuario>()), Times.Once);
         }
 
         [Fact]
-        public void FindAll_ReturnsListOfUsuarioVM()
+        public void FindAll_Should_Returns_List_Of_UsuarioVM()
         {
+            // Arrange         
             var usuario = _usuarios.First();
             usuario.PerfilUsuario = PerfilUsuario.Administrador;
             usuario.StatusUsuario= StatusUsuario.Ativo;
-            var idUsuario = usuario.Id;
+            var idUsuario = usuario.Id;            
             
-            // Arrange         
             _repositorioMock.Setup(repo => repo.GetAll()).Returns(_usuarios);
             _repositorioMock.Setup(repo => repo.Get(idUsuario)).Returns(usuario);
 
@@ -53,18 +53,20 @@ namespace Test.XUnit.Business.Implementations
             // Assert
             Assert.NotNull(result);            
             Assert.IsType<List<UsuarioVM>>(result);
-            Assert.Equal(_usuarios.Count, result.Count);           
+            Assert.Equal(_usuarios.Count, result.Count);
+            _repositorioMock.Verify(repo => repo.GetAll(), Times.Once);
+            _repositorioMock.Verify(repo => repo.Get(idUsuario), Times.Once);
         }
 
         [Fact]
-        public void FindAll_ReturnsNull()
+        public void FindAll_Should_Returns_Null()
         {
+            // Arrange         
             var usuario = _usuarios.First();
             usuario.PerfilUsuario = PerfilUsuario.Usuario;
             usuario.StatusUsuario = StatusUsuario.Ativo;
             var idUsuario = usuario.Id;
-
-            // Arrange         
+            
             _repositorioMock.Setup(repo => repo.GetAll()).Returns(_usuarios);
             _repositorioMock.Setup(repo => repo.Get(idUsuario)).Returns(usuario);
 
@@ -73,32 +75,33 @@ namespace Test.XUnit.Business.Implementations
 
             // Assert
             Assert.Null(result);
+            _repositorioMock.Verify(repo => repo.GetAll(), Times.Never);
+            _repositorioMock.Verify(repo => repo.Get(idUsuario), Times.Once);
         }
 
         [Fact]
-        public void FindById_ReturnsParsedUsuarioVM()
+        public void FindById_Should_Returns_Parsed_UsuarioVM()
         {
             // Arrange
             var usuario = _usuarios.First();
-            var id = usuario.Id;
-            
+            var idUsuario = usuario.Id;            
 
-            _repositorioMock.Setup(repo => repo.Get(id)).Returns(usuario);
+            _repositorioMock.Setup(repo => repo.Get(idUsuario)).Returns(usuario);
 
             // Act
-            var result = _usuarioBusiness.FindById(id);
+            var result = _usuarioBusiness.FindById(idUsuario);
 
             // Assert
             Assert.NotNull(result);
             Assert.IsType<UsuarioVM>(result);
-            Assert.Equal(usuario.Id, result.Id);            
+            Assert.Equal(usuario.Id, result.Id);
+            _repositorioMock.Verify(repo => repo.Get(idUsuario), Times.Once);
         }
 
         [Fact]
-        public void Update_ReturnsParsedUsuarioVM()
+        public void Update_Should_Returns_Parsed_UsuarioVM()
         {
-            // Arrange
-            
+            // Arrange            
             var usuario = _usuarios.First();
             var usuarioVM = new UsuarioMap().Parse(usuario);
             usuario.Nome = "Teste Usuario Update";                       
@@ -112,10 +115,11 @@ namespace Test.XUnit.Business.Implementations
             Assert.NotNull(result);
             Assert.IsType<UsuarioVM>(result);
             Assert.Equal(usuarioVM.Id, result.Id);
+            _repositorioMock.Verify(repo => repo.Update(It.IsAny<Usuario>()), Times.Once);
         }
 
         [Fact]
-        public void Delete_ReturnsTrue()
+        public void Delete_Should_Returns_True()
         {
             // Arrange
             var obj = new UsuarioMap().Parse(_usuarios.First());
@@ -125,7 +129,9 @@ namespace Test.XUnit.Business.Implementations
             var result = _usuarioBusiness.Delete(obj);
 
             // Assert
+            Assert.NotNull(result);
             Assert.True(result);
+            _repositorioMock.Verify(repo => repo.Delete(It.IsAny<Usuario>()), Times.Once);
         }
     }
 }
