+++ despesas-backend-api-net-core.XUnit/Domain/Entities/LancamentoTest.cs
@@ -6,7 +6,7 @@
         [InlineData(1, 100.6, "Descrição 1", 1, typeof(Usuario), 1, typeof(Despesa), 1 , typeof(Receita), 1, typeof(Categoria))]
         [InlineData(2, 58.98, "Descrição 2", 2, typeof(Usuario), 3, typeof(Despesa), 2, typeof(Receita), 2, typeof(Categoria))]
         [InlineData(3, 5000, "Descrição 2", 3, typeof(Usuario), 4, typeof(Despesa), 6, typeof(Receita), 54, typeof(Categoria))]
-        public void Lancamento_ShouldSetPropertiesCorrectly(int id, Decimal valor, string descricao, int usuarioId, Type typeUsuario, int despesaId, Type typeDespesa, int receitaId, Type typeReceita,  int categoriaId, Type typeCategoria)
+        public void Lancamento_Should_Set_Properties_Correctly(int id, Decimal valor, string descricao, int usuarioId, Type typeUsuario, int despesaId, Type typeDespesa, int receitaId, Type typeReceita,  int categoriaId, Type typeCategoria)
         {
             var mockUsuario = Mock.Of<Usuario>();
             var mockDespesa = Mock.Of<Despesa>();
