+++ despesas-backend-api-net-core.XUnit/Domain/VM/CategoriaVMTest.cs
@@ -7,7 +7,7 @@
         [InlineData(1, "Test Categoria Description 1", 1, TipoCategoria.Todas)]
         [InlineData(2, "Test Categoria Description 2", 2, TipoCategoria.Despesa)]
         [InlineData(3, "Test Categoria Description 3", 3, TipoCategoria.Receita)]
-        public void CategoriaVM_ShouldSetPropertiesCorrectly(int id, string descricao, int idUsuario, TipoCategoria tipoCategoria)
+        public void CategoriaVM_Should_Set_Properties_Correctly(int id, string descricao, int idUsuario, TipoCategoria tipoCategoria)
         {
             // Arrange
             var categoriaVM = new CategoriaVM();
