+++ despesas-backend-api-net-core.XUnit/Domain/VM/LoginVMTest.cs
@@ -6,7 +6,7 @@
         [InlineData(1, "user1@user1.com", "userTeste1", "userTeste1")]
         [InlineData(2, "user2@user2.com", "userTeste2", "userTeste2")]
         [InlineData(3, "user3@user3.com", "userTeste3", "userTeste3")]
-        public void LoginVM_ShouldSetPropertiesCorrectly(int idUsuario, string email, string senha, string confirmaSenha)
+        public void LoginVM_Should_Set_Properties_Correctly(int idUsuario, string email, string senha, string confirmaSenha)
         {
             // Arrange and Act
             var loginVM = new LoginVM
