--- despesas-backend-api-net-core.XUnit/Infrastructure/Data/Repositories/Generic/GenericRepositorioCategoriaTest.cs
@@ -1,176 +1,0 @@
-﻿using despesas_backend_api_net_core.Infrastructure.Data.Repositories.Generic;
-using Xunit.Extensions.Ordering;
-
-namespace Test.XUnit.Infrastructure.Data.Repositories.Generic
-{
-    public class GenericRepositorioCategoriaTest
-    {
-        private Mock<RegisterContext> _dbContextMock;
-        private Mock<GenericRepositorio<Categoria>> _repository;
-
-        public GenericRepositorioCategoriaTest()
-        {
-            // Arrange
-            var options = new DbContextOptionsBuilder<RegisterContext>()
-                .UseInMemoryDatabase(databaseName: "TestDatabase")
-                .Options;
-
-            _dbContextMock = new Mock<RegisterContext>(options);
-            _dbContextMock.Setup(c => c.Set<List<Categoria>>());
-            _repository = new Mock<GenericRepositorio<Categoria>>(_dbContextMock);
-        }
-
-        [Fact]
-        public void Insert_ShouldAddItemToDataSetAndSaveChanges()
-        {
-            // Arrange
-            var item = new Categoria();
-            var dataSet = new List<Categoria>();
-            var dbSetMock = Usings.MockDbSet(dataSet);
-            _dbContextMock.Setup(c => c.Set<Categoria>()).Returns(dbSetMock.Object);
-            var repository = new GenericRepositorio<Categoria>(_dbContextMock.Object);
-
-            // Act
-            var result = repository.Insert(item);
-
-            // Assert
-            Assert.Equal(1, dataSet.Count);
-            Assert.Contains(item, dataSet);
-            _dbContextMock.Verify(c => c.SaveChanges(), Times.Once);
-            Assert.Equal(item, result);
-        }
-
-        [Fact]
-        public void GetAll_ShouldReturnAllItemsFromDataSet()
-        {
-            // Arrange
-            var items = CategoriaFaker.Categorias();
-            var dataSet = items;
-            var dbSetMock = Usings.MockDbSet(dataSet);
-            _dbContextMock.Setup(c => c.Set<Categoria>()).Returns(dbSetMock.Object);
-            var repository = new GenericRepositorio<Categoria>(_dbContextMock.Object);
-
-            // Act
-            var result = repository.GetAll();
-
-            // Assert
-            Assert.Equal(items.Count, result.Count);
-            Assert.Equal(items, result);
-        }
-
-        [Fact, Order(1)]
-        public void Get_ShouldReturnItemWithMatchingId()
-        {
-            // Arrange
-            var itens = UsuarioFaker.Usuarios();
-            var item = itens.First();
-            var itemId = item.Id;
-            
-            var dataSet = itens;
-            var dbSetMock = Usings.MockDbSet(dataSet);
-            _dbContextMock.Setup(c => c.Set<Usuario>()).Returns(dbSetMock.Object);
-            var repository = new GenericRepositorio<Usuario>(_dbContextMock.Object);
-
-            // Act
-            var result = repository.Get(itemId);
-
-            // Assert
-            Assert.NotNull(result);
-            Assert.IsType<Usuario>(result);
-            Assert.Equal(item, result);
-        }
-
-        [Fact]
-        public void Update_ShouldUpdateItemAndSaveChanges()
-        {
-            // Arrange
-            var dataSet = CategoriaFaker.Categorias();
-            var existingItem = dataSet.First();
-            var dbContext = new RegisterContext(new DbContextOptionsBuilder<RegisterContext>()
-                .UseInMemoryDatabase(databaseName: "TestDatabase")
-                .Options);
-
-            var repository = new GenericRepositorio<Categoria>(dbContext);
-
-            // Act
-            var result = repository.Insert(existingItem);
-            var updatedItem = new Categoria
-            {
-                Id = result.Id,
-                Descricao = "Teste Update Item",
-                UsuarioId = result.Id,
-                Usuario = result.Usuario,
-                TipoCategoria = TipoCategoria.Receita
-            };
-
-            result = repository.Update(updatedItem);
-
-            // Assert
-            Assert.NotNull(result);
-            Assert.NotEqual(existingItem, result);
-            Assert.Equal(updatedItem.Descricao, result.Descricao);
-            Assert.Equal(updatedItem, result);            
-        }
-
-
-        [Fact]
-        public void Update_Should_UpdateItem_And_ReturnNull()
-        {
-            // Arrange
-            var dataSet = CategoriaFaker.Categorias();
-            var existingItem = dataSet.First();
-                   
-            var dbContext = new RegisterContext(new DbContextOptionsBuilder<RegisterContext>()
-                .UseInMemoryDatabase(databaseName: "TestDatabase")
-                .Options);
-
-
-            var repository = new GenericRepositorio<Categoria>(dbContext);
-
-            // Act
-            var result = repository.Update(existingItem);
-
-            // Assert
-            Assert.Null(result);
-        }
-
-
-        [Fact]
-        public void Delete_With_Existing_Item_Should_Remove_Item_And_SaveChanges()
-        {
-            // Arrange
-            var lstCategorias = CategoriaFaker.Categorias();
-            var item = lstCategorias.Last();
-            var dataSet = lstCategorias;
-
-            var dbSetMock = Usings.MockDbSet(dataSet);
-            _dbContextMock.Setup(c => c.Set<Categoria>()).Returns(dbSetMock.Object);
-            var repository = new GenericRepositorio<Categoria>(_dbContextMock.Object);
-
-            // Act
-            var result = repository.Delete(item);
-
-            // Assert
-            Assert.True(result);
-            _dbContextMock.Verify(c => c.SaveChanges(), Times.Once);
-        }
-
-        [Fact]
-        public void Delete_WithNonExistingItem_ShouldNotRemoveItemAndReturnFalse()
-        {
-            // Arrange
-            var dataSet = CategoriaFaker.Categorias();
-            var item = new Categoria { Id = 0 };
-            var dbSetMock = Usings.MockDbSet(dataSet);
-            _dbContextMock.Setup(c => c.Set<Categoria>()).Returns(dbSetMock.Object);
-            var repository = new GenericRepositorio<Categoria>(_dbContextMock.Object);
-
-            // Act
-            var result = repository.Delete(item);
-
-            // Assert            
-            Assert.False(result);
-            _dbContextMock.Verify(c => c.SaveChanges(), Times.Never);
-        }
-    }
-}
